remove SubsetRequest as it's expected to be useless
